requirements.txt
google-cloud-vision==1.0.0